# PERF: improved performance of multiindex slicing