closes #10287
 
master 
```
In [22]: %timeit mdt2.loc[idx[test_A-eps_A:test_A+eps_A],:].loc[idx[:,test_B-eps_B:test_B+eps_B],:].loc[idx[:,:,test_C-eps_C:test_C+eps_C],:].loc[idx[:,:,:,test_D-eps_D:test_D+eps_D],:]
10 loops, best of 3: 141 ms per loop

In [23]: %timeit mdt2.loc[idx[test_A-eps_A:test_A+eps_A,test_B-eps_B:test_B+eps_B,test_C-eps_C:test_C+eps_C,test_D-eps_D:test_D+eps_D],:]
1 loops, best of 3: 4.23 s per loop
```

PR
```
# this actually is a bit faster in master. but this repeated chain indexing is frowned upon anyhow
In [22]: %timeit mdt2.loc[idx[test_A-eps_A:test_A+eps_A],:].loc[idx[:,test_B-eps_B:test_B+eps_B],:].loc[idx[:,:,test_C-eps_C:test_C+eps_C],:].loc[idx[:,:,:,test_D-eps_D:test_D+eps_D],:]
1 loops, best of 3: 210 ms per loop

# this is the prefered method (as you can set with this and such), and such a huge diff.
In [23]: %timeit mdt2.loc[idx[test_A-eps_A:test_A+eps_A,test_B-eps_B:test_B+eps_B,test_C-eps_C:test_C+eps_C,test_D-eps_D:test_D+eps_D],:]
1 loops, best of 3: 425 ms per loop
```

master
```
In [10]: %timeit s3.isin([1,2])
100 loops, best of 3: 8.83 ms per loop
```
 
PR
```
In [2]: s3 = Series(np.random.randint(1, 10, 100000)).astype('int64')

In [5]: %timeit s3.isin([1,2])
100 loops, best of 3: 2.47 ms per loop

```