@jreback 2015-06-05
@sinhrks IIRC we had some discussion of the specialized ``ismember_int64``. added here.