@jreback 2015-06-07
going to defer to 0.17.0 as 3.2 fails for some reason (and not going thru the trouble of creating a build environment); we are going to drop in #9118 . Its pretty hard to install 3.2.